sdk/requirements.txt
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.1
azure-mgmt-cosmosdb>=9.3.0